+++ tech1-framework-b2b-base-security-jwt/src/main/java/io/tech1/framework/b2b/base/security/jwt/resources/BaseSuperAdminResource.java
@@ -69,6 +69,11 @@ public class BaseSuperAdminResource {
         return this.baseSuperAdminService.getSessions(cookie);
     }
 
+    @PostMapping("/sessions/{sessionId}/renew/manually")
+    public void renewManually(@PathVariable UserSessionId sessionId, HttpServletRequest httpServletRequest) {
+        this.baseUsersSessionsService.renewUserRequestMetadataManually(sessionId, httpServletRequest);
+    }
+
     @DeleteMapping("/sessions/{sessionId}")
     @ResponseStatus(HttpStatus.OK)
     public void deleteById(@PathVariable UserSessionId sessionId) {