+++ RELEASE_NOTES.md
@@ -6,7 +6,8 @@
 - Add "create_at", "updated_at", "metadata_renew_cron", "metadata_renew_manually" → tech1_users_sessions (MongoDB) + MIGRATION
 - Add "enableSessionsMetadataRenewCron" configuration
 - Add properties 2.0: @AllArgsConstructor(onConstructor = @__({@ConstructorBinding}))
-- Add /renew/manually/{sessionId} endpoint + Postman
+- Add /sessions/manually/{sessionId} endpoint + Postman
+- Add /superadmin/sessions/{sessionId}/renew/manually endpoint + Postman
 
 ### v1.15
 - Move <distributionManagement> ossrh → "release" profile